Remove the start accepting / stop accepting transitions
This makes the design generally simpler and re-uses a bit more of the systemd
semantic.

With this change the server can function normally without having a
crankized/non-crankized version. FDs are listened to if LISTEN_FDS is passed
and status updates are sent if NOTIFY_FD is passed. Notification updates are
otherwise noop operations.

Crank has less state changes to track which will make it also easier to
implement with all the possible corner-cases.

Use a socketpair of type SOCK_DGRAM to make sure messages are send with a good
size (and because it's what sd-notify uses, but they use a name socket).